Move substitute tests to PodTemplateUtilsTest
