+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/PodTemplateUtilsTest.java
@@ -30,6 +30,9 @@ import static org.csanchez.jenkins.plugins.kubernetes.PodTemplateUtils.*;
 import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.*;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import org.csanchez.jenkins.plugins.kubernetes.model.KeyValueEnvVar;
 import org.csanchez.jenkins.plugins.kubernetes.model.SecretEnvVar;
 import org.junit.Test;
@@ -325,5 +328,45 @@ public class PodTemplateUtilsTest {
         ContainerTemplate result = combine(template1, template2);
 
         assertThat(result.getEnvVars(), empty());
+    }
+
+    // Substitute tests
+
+    @Test
+    public void shouldIgnoreMissingProperties() {
+        Map<String, String> properties = new HashMap<>();
+        properties.put("key1", "value1");
+        assertEquals("${key2}", substitute("${key2}", properties));
+    }
+
+    @Test
+    public void shouldSubstituteSingleEnvVar() {
+        Map<String, String> properties = new HashMap<>();
+        properties.put("key1", "value1");
+        assertEquals("value1", substitute("${key1}", properties));
+    }
+
+    @Test
+    public void shouldSubstituteMultipleEnvVars() {
+        Map<String, String> properties = new HashMap<>();
+        properties.put("key1", "value1");
+        properties.put("key2", "value2");
+        assertEquals("value1 or value2", substitute("${key1} or ${key2}", properties));
+    }
+
+    @Test
+    public void shouldSubstituteMultipleEnvVarsAndIgnoreMissing() {
+        Map<String, String> properties = new HashMap<>();
+        properties.put("key1", "value1");
+        properties.put("key2", "value2");
+        assertEquals("value1 or value2 or ${key3}", substitute("${key1} or ${key2} or ${key3}", properties));
+    }
+
+    @Test
+    public void shouldSubstituteMultipleEnvVarsAndUseDefaultsForMissing() {
+        Map<String, String> properties = new HashMap<>();
+        properties.put("key1", "value1");
+        properties.put("key2", "value2");
+        assertEquals("value1 or value2 or defaultValue", substitute("${key1} or ${key2} or ${key3}", properties, "defaultValue"));
     }
 }
